Commit many new changes to osm2pgsql, including: - list of tags read from file - slim mode works again - relations properly supported - more efficient DB usage
It includes some restructuring of the code, in particular, the output module
manages the mid-level now and the main program doesn't call it at all. This
moves many of the previous hacks to the output module which can manage the
mid-level as appropriate for its output (i.e. slim mode requires different
semantics from ram mode)


git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@7331 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4