Adds tile expiry support - see http://lists.openstreetmap.org/pipermail/dev/2009-February/013934.html
This introduces 2 new commandline 
options: "-e <zoomlevel>" and "-o <dirty tile list output file>".  So, 
specifying "-e 17 -o /tmp/dirty_tiles" when importing a delta will cause 
osm2pgsql to generate a list of all zoom level 17 tiles which the delta 
has made dirty and store it in /tmp/dirty_tiles.

Proviso: for polygons, it currently takes a simplistic approach of drawing 
a bounding box around the whole polygon and marking every tile in the box 
as dirty.  If the bounding box is large (over 30x30Km) the polygon is 
treated as a line instead, so only the perimeter will be marked as dirty 
(this is so that huge polygons don't expire vast numbers of tiles and is 
based on the assumption that we probably aren't going to shade the area 
of massive polygons).

The dirty tile list is maintained in memory as a binary tree and dumped to 
disk at the end of the run.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@13612 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4