Fix bug in hstore rework.
Bad idea to remove the k/v pairs from the linked list after
writing their colum, as write_wkts is called more than once.

So flag them to have their own column.

Afterwords check this flag to see if a given
k/v pair should be written to the hstore or not.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@25786 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4