[Osm2pgsql] Add an option to disable parallel indexing
This is a patch by Hartmut Holzgraefe ( https://github.com/hholzgra/osm2pgsql/commit/f779298ba9b2cc84b3949f3fd4b684ea64bcee34),
although it was adapted to have parallel indexing the default and the option disables it.

Under the conditions I have tested, parallel indexing seemed to be slightly faster than sequential indexing, but that may depend
on hard and or software, as well as the size of the planet (extract) to import, hence the option to disable it. 




git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@26664 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4