[osm2pgsql] allocate node cache in one big chunk instead of small pieces
This is a patch by Hartmut Holzgraefe ( https://github.com/hholzgra/osm2pgsql/commit/9f05f1b925751eca657bbd7fa6cbac937f7756d2 ),
adapted to add a command-line switch to revert to the old behavior

Allocating the node cache in small chunks causes the memory not to be returned to the operating system after it is free()ed,
as it is using a different internal allocation strategy (not mmap based).

With the new (default) behavior the entire node cache is allocated in one single malloc call.

At least on Linux, physical memory is still only allocated in chunks when it is needed and only virtual memory is reserved in one chunk.
So apart from the correct freeing of memory during the indexing stage, behavior doesn't change.
Other operating systems might behave differently though, which is why the old strategy can still be selected via a command line switch.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@26665 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4