[osm2pgsql] Parralellize pending ways / relations
Process the stage of "going over pending ways / relations" in parallel,
by spawning multiple worker processes that each work through part of the list
of pending ways / relations independent of each other. The list is staticly partitioned
and assigned to the workers.

The number of processes spawned can be set via a commandline option