[osm2pgsql] Parralellize pending ways / relations
Process the stage of "going over pending ways / relations" in parallel,
by spawning multiple worker processes that each work through part of the list
of pending ways / relations independent of each other. The list is staticly partitioned
and assigned to the workers.

The number of processes spawned can be set via a commandline option



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@26893 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4