[osm2pgsql] set synchronous_commit to off
The introduction of the parallelisation during the "going over pending ways" phase
required to not run this phase in an extended transaction, as it would otherwise
deadlock. As now each pending way is processed in a separate transaction, the commit
rate of postgresql became critical for performance. If run with default postgresql parameters
of fsync=on synchronous_commit=on, it resulted in a massive performance degredation.

See report on dev ( http://lists.openstreetmap.org/pipermail/dev/2011-December/023843.html )

Setting the parameter synchronous_commit to off from within postgresql can recover most of this performance.

As all data is derived data from the main osm data, the potential loss of transactions during a crash of the database is
acceptable. Full imports will need to be restarted from scratch anyway, and diff imports should recover on the next diff import.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@27179 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4