[osm2pgsql] Add a simple testing script to run through most options
In order to help prevent bugs in developing osm2pgsql, add a script that automatically runs through a number of tests
to exercise a diverse set of parameters.

In this version it simply runs osm2pgsql multiple times with different sets of parameters each time. It can
therefore currently only detect errors where osm2pgsql fails, but not where it creates wrong output. In future
additional checks on the output might be added though.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@28333 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4