Implement a flat file persistent node cache
This implements a flat file persistent node cache for slim mode. It replaces the use of pgsql for node caching.
It stores lat / lon pairs in a large monolithic file ordered by osm id. This allows access to nodes in O(1). Assuming a dense
use of osm_id's (for the full planet this is currently at ~80%), this is the most efficient way to represent the data.

The planet_osm_nodes table with indexes currently takes about 100GB for the full planet. In contrast the flat persistent node
cache file is < 16GB. This should allow the disk caches to be more efficient. Furthermore, each read will load in more useful data.
Both aspects should reduce the number of seeks necessary on disk to retrieve node information.

Therefore hopefully this method should be faster (although I don't yet have benchmarks for this) than storing the data in pgsql.
Particularly diff imports may benefit, as during diff imports typically a low hit rate on the ram cache is achieved.

As the density of valid nodes in the osm id range reduces, e.g. due to only importing extracts, the effectivness of this
method quickly reduces. Therefore it is not well suited for extracts.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@28513 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4