Use async IO in persistent_cache_nodes_get_list
Issue multiple IO requests for all of the blocks in the node list simultaniously (via async IO).
This should speed up processing as higher queue depths improve throughput.

This uses the GNU C library aio functions. In its current implementation of the library, this is not
"true" async IO at the kernel level, but instead the library implements a thread pool to handle multiple
simultanious requests.



git-svn-id: http://svn.openstreetmap.org/applications/utils/export/osm2pgsql@28514 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4