Add support for Czech/Slovak house number tagging scheme for Nominatim.
In CZ and SK, a house number is composed of two numbers: a conscription
number and a street number. The conscription number is unique within the
place (city, town, village), and the street number is local to the street
the house is located at. Usually this is written as 1135/45, where 1135 is
the conscription number, and 45 is the street number. In small villages
houses don't usually have individual street numbers.

The patch enables osm2pgsql to read addr:conscriptionnumber and
addr:streetnumber tags used in CZ and SK, in Gazetteer mode. If any of those
two is found alone, it supersedes addr:housenumber, if both exist, they are
joined with a slash, and used instead addr:housenumber as well. This is done
because addr:housenumber is often used to specify what Mapnik would render (as
it doesn't currently recognise additional tags), so quite often it contains
incomplete information.

--HG--
extra : rebase_source : b21c6e411f627738d20e9a7baf9299c1add6fe6d