Add a new regression test suite using integration tests
This regression test suite imports well defined osm test files
(currently a liechtenstein snapshot for general testing and
 a special crafted osm file with many different multi-polygon examples )
with various different parameters of osm2pgsql. It then runs various
different SQL queries against the resulting db to verify the results are
as expected.

Although the framework is in place, the number of tests are still rather limited
and need expanding to cover all of the interesting cases.