Improve tag copying of outer way tags to multi-polygon relations
Previously outer way tags were only used to determine the multi-polygon
relation if the relation was untagged or only tagged with a name.

With hstore, tags like "fixme" could disrupt this logic, as relations
were no longer untagged and therefore didn't inherit the outer way tags.

Osm2pgsql now only looks to see if the relation is "untagged" with respect
to tags determining if something is a polygon. I.e. any tags that are marked
"polygon" in the style sheet. Other tags like name or fixme are ignored