Fix diff bug that resulted in missing polygon after relation deleting
When an outer way is part of a relation, it only appears in the rendering
tables as part of the relation. However, when the relation gets deleted,
the now individual way needs to get re-added into the rendering tables.
This did not happen previously.

This fixes #67 which is a duplicate of trac issue 4525.