Add a new minimal style for hstore or a starting point
This style will not create any tag columns and is designed for use with hstore.

It also deletes various unwanted tags from imports to save space.

The approximate rule used for deciding to add a tag was if the
import had >1m tags to drop and the tag had >100k.