Add tests for multi-polygon diff import issues reported in trac #2853
When using old-style multi-polygon tagging with both outer and inner ways
tagged with the same tags, then if the tags on the outer ways change,
osm2pgsql should create a new inner polygon to be consistent with fresh
imports.

This is currently not the case, so osm2pgsql doesn't pass these tests. They
are for now therefore disabled