Fix deduplicate_strings was not initialised.
In non-slim mode, all tags are stored in memory for the entire import
time. There are a lot of duplicate strings in tags, like highway=residential.
Instead of using up 18 characters to store the entire string each time, simple
use pointers to a single deduplicated version of the tag.

In slim mode however, tag strings are only stored for a short time in memory
before getting written to the database. So there is no need to add the extra
effort and thread contention of de-duplicating strings.