Fix style.lua error with non numerical layer values
The value for the layer tag should be numeric. But there are a few non numeric values in the OSM db - http://taginfo.openstreetmap.org/keys/layer#values This causes an error of something like: "attempt to perform arithmetic on local 'layer' (a string value)". Thus we check to see if the value is a number.

All credits for this go to @systemed