Started adding unit tests.
Added a very basic XML parsing unit test, which is run in the
standard autoconf manner. Also ported the existing regression
tests to be run from the autoconf test harness as well.

Part of this meant pulling most of the code into a temporary
library so that it can be re-used by all the binaries, including
the unit tests. Hopefully this won't cause any major issues - it's
all libtool-based, so should be portable, and is statically linked
to the resulting binary so needs no installation.