Added first pass of moving 'done' logic to output.
The middle keeps track of the items which have changed due to
updates to the OSM elements, i.e: a way is marked pending because
a node that it uses has changed. However, the output also made
some decisions about what to track by passing a `pending` flag to
some methods, and calling `ways_done` or `relations_done` for some
elements.

We want to have multiple outputs for the same middle, so we need
these decisions to be made for each individual output. This commit
is the first step, by moving all the custom 'done' tracking to the
output types via a shared, provisional tracker, which will be
written properly in the next pass.