Abstract `expire_tiles` output.
Rather than just outputting to a file at destructor time, which
has the potential to go horribly wrong if anything is thrown, this
now needs `output_and_destroy()` to be called before the object
goes out of scope.

The output is now effectively a functor object, so this allows it
to be swapped out for testing purposes.