Use GeoHash ordering instead of way ordering
Creating a table as ORDER BY way is known to offer no performance
advantages (#87) and in fact ahve losses in some cases. GeoHash
offers a better way to have geograhically nearby data in the same
or nearby pages.

Rather than simply doing an order on ST_GeoHash(ST_Transform(way,4326)),
we can get a 9% performance gain by only transforming the ST_Envelope.

This works in all cases on EPGS 3857 as geometries cannot cross WGS84
boundaries without also crossing 3857 boundaries.

There might be difficulties in projections that cross the 180 line such
as Alaskan planes, but as OSM data is itself WGS84 any geometries will have
to be broken at the 180 line anyways, and it remains safe. The worst
potential consequence would be a slightly unoptimal ordering on-disk.

Another 5% performance can be gained by restricting the geohash length to
10 characters. This gives 25 bits of resolution in either direction, which
is sufficient for our purposes.

Fixes #208