move static buffers into middle-pgsql_t class
Replaces the two static buffers for query building by one huge buffer
and changes the buildier functions such that they fill a povided buffer
insted of providing the buffer themselves. This avoids a lot of allocations
in the import case (when copy is used) but increases the number for
updates (when insert is used).