Add a highway example to the sample multi style
This gives a demonstration of points, lines, and areas. It also
shows how multi-lingual names for rendering can be computed in
advance, using a lookup table for object selection, and using a
lookup table for tag conversion.

Unfortunately, it exposes some bugs in osm2pgsql - but the
conversion itself is fine.