Allow polygon/linestring decision to be set by value in lua
This allows the decision whether an object is treated as polygon or linestring
to be set in the style.lua file.

In particular, it treats man_made=embankment and natural=cliff as linestring,
and highway=services as polygon.

Including this in the default style.lua file has the following advantages:
* It serves as an easy-to-extend example of how the polygon/linestring decision
  can be made based on key.
* It provides sane defaults - it can be expected that most users of the style
  want to treat man_made=embankment and natural=cliff as linestring, and
  highway=services as polygon.

This resolves the following downstream bugs in openstreetmap-carto:
* https://github.com/gravitystorm/openstreetmap-carto/issues/892
* https://github.com/gravitystorm/openstreetmap-carto/issues/268
* https://github.com/gravitystorm/openstreetmap-carto/issues/137

This builds on #345, and thus assumes #345 is merged first.