better separation of read and write cache in flatnode store
The read and write cache blocks are different caches of the same
file and must not be used at the same time. Furthermore, the read
cache must not be written, when multiple threads are used.

The logic for the two caches has been cleanly separated now
and the two conditions are enforced via asserts.