Increase dense block size from 1k to 8k nodes
The dense cache has a fixed memory cost which is associated with
the number of blocks needed. Increase the block size reduces the
number of blocks, resulting in reduced memory usage. For this change
on modern hardware it is a decrease from 1GB to 128MB. Large blocks
are less efficient than small ones, but it turns out that this does
not matter for planet imports, where the efficiency change is small
and memory usage is actually slightly reduced, or for small imports,
where the fixed cost change overwhelms anything else.

On large but not planet sized extracts (e.g. Europe) memory usage
is slightly increased, by about 450MB, while the cache in total
consumes about 18GB.

Changing the cache size did not have a measurable impact on
performance in testing.

Fixes #439