Remove check for unlogged and PostgreSQL < 9.1
PostgreSQL 9.0 is 5 years old and will not have any new releases as
it is unsupported. This change doesn't remove support for 9.0, but
just the check with --unlogged that the server supports it.

If someone does try --unlogged with an older version, it will error
early in the process when it tries to create tables.