add flag to process new-style polygons only
With this flag set, the logic that checks for tags of outer
rings on multipolygons is disabled and such ways also cannot
be superseded and will always be rendered. Note that for
efficiency reasons we still process polygon ways in the
second stage where we can make use of multiple threads.