use libosmium types in middle layer
Passes the original libosmium buffers into middle classes
when storing the data. This also moves the storing of additional
attributes to the middle layer. Data is still retrieved in the
old osm2pgsql formats.

Requires a rewrite of the test for the middle layer. Now
uses a libosmium buffer to create temporary objects. Cleaned
up code a bit on that occasion.