move middle::relations_get() to libosmium types
Relations are now created in an osmium buffer which needs to be
provided by the caller. The output modules now have a per-instance
buffer for that.

middle-ram has to go some hoops to convert its saved member list.
(see for_builder() function in memberlist_t) but in the long run
it should save the data in libosmium format anyways.