don't pre-filter relation member tags
Tags for ways of multipolygons are no longer prefiltered
before handing them into the relation member filter function.
For new-style MPs the tags are not relevant anyways. For
old-style MPs, the filter now first creates a list of common
tags and then filters the result with the standard relation
tags filter.

Fixes #605.