uint32_t for zoom level in options_t, zoom level 0 for "no tile expiry"
The constructor of expire_tiles_t expects uint32_t and the conversion to
unsigned has to happen anywhere. In addition, some more checks on the
arguments supplied by the user and a unit test for option `-e` does not harm.

This commit adapts output_pgsql_t and output_multi_t because zoom level 0
for the tile expiry now means that no expiry output is requested.