refactor getting read-only middle object
Rewrite get_instance() to receive the shared pointer as a parameter
and get rid of the no_Delete() hack which essentially renders the
shared pointer mute.

Also get rid of const for middle_query_t. There is no point in
adding this as the middle_query_t interface is completely const
already.