Introduce flush when type of import object changes
This replaces the endCopy operation in the middle_t
read functions. These functions are const and never
should have called a function that changes state.
Instead of falling out of copy then, all copy operations
are finished, when processing of an OSM type is done.
Read functions should never be called when actually
processing the same table, so the one-time end-copy
should be sufficient.