split middle-pgsql into insert and query class
The query instance can work on a single connection. This saves
a couple of connections in the processing stage.

Requires slightly changing the behviour of the connections
that copy in the data. They need to commit, so that the
data becomes visible.

middle also needs to start the tables earlier, because
the tables must have been created when preparing queries
in the query connection.