gazetteer: handle updates of unnamed places
When downgrading an unnamed object from SF_MAIN to SF_MAIN_NAMED,
the object needs to be deleted. Finding that condition in has_data()
is a bit expecsive, so instead catch the condition when copying
the data out. If there are multiple main tags, then
delete_unused_classes() needs to remove the instance as well.
Adapt has_place() for that.

Fixes #933.