Reduce max number of default parallel threads to 4
Parallel processing is used only when indexes are created and
in the second processing stage of updates. In both cases the
number of parallel threads should not be too high. Indexing
requires a lot of memory and having too many in parallel running
actually slows down the process. Updates are generally run in
parallel with other load (like rendering) and should therefore
also not take up too many CPUs.

Fixes #885.