Move declaration of static `db` into tests
`pg::tempdb_t db` or `testing::db::import_t db` is defined as static in
many test files. This is problematic, because the class may throw errors
which we can't catch and the testing framework can't report correctly
then.

This is caught by clang-tidy with error messages like this:
"warning: initialization of 'db' with static storage duration may throw
an exception that cannot be caught [cert-err58-cpp]"

This commit fixes all these cases.

This commit also adds a bunch of `const`s.