Make sure PostgreSQL connection is cleaned up properly.
PQfinish(m_conn) must also be called if the connection attempt fails.
By wrapping this into a unique_ptr with deleter, the code becomes
simpler, none of the special member functions is needed any more.

This also adds a little convenience function error_msg() which makes the
rest of the code somewhat cleaner.