gazetteer: move main tag filtering into processing function
Filtering of main tags (i.e. handling of named-only tags or
fallback tags) must happen before the actual output because
the filtered list is already needed when computing the list
of classes to keep. Currently the function that creates the
class list does its own filtering and is inconsisten with
what the output function does.