Limit the number of scheduled deletes
As we have to issue a delete for each and every untagged node,
the number of scheduled deletes can become quite large even for
a small change file. It is better to send off the copy buffer
a bit earlier in this case to ensure that the deletes can already
run in parallel with the parsing of the file. The change also
ensures that the memory usage of the deleter is bound.