Make pg_result_t into a real type and add convenience functions
The convenience functions come from the result_t type used only in
tests which isn't necessary any more and is removed. This removes
code duplication and makes sure we are testing more of our own
code.

The extra convenience functions that are not used in the production code
should be removed by the compiler because they are all defined inline
in the class.