Bugfix: Threading problem in db_copy_thread_t
The finish_copy() method was called from the main program and from the
copy thread leading to a race condition.

I have fixed not only this but clearly separated out everything
happening in the main thread from everything in the copy thread by
putting the code for the latter into its own class. This way a
problem like this can not happen again.