Allow decision on single/multi geometry for each geometry
With the old code the decision whether to generate many single polygons
or one multipolygon is done when creating the osmium_builder_t object
(based on a command line argument). This change allows this to be set
when each (multi)polgyon is build. This flexibility will be used later
to allow setting this on a table-by-table basis.