Imported newest version of libosmium 2.15.5
Also the code was moved from contrib/libosmium/ to
contrib/libosmium/include to be more consistent with the other vendored
in code.

In addition this changes the minimum required libosmium version 2.15.4
which contains an important fix related to multipolygon building.
Versions of libosmium before that can use enormous amounts of memory
when trying to build special complex multipolygons.