Simplify expire tree merging
This also fixes a (theoretical) bug introduced a bit earlier when clones
were only generated for outputs that need forward depencencies. In that
case the outputs and clones used in merging expire trees did not match.
The bug is theoretical, because there is currently no way in osm2pgsql
to create several outputs of which some need dependency management and
others don't.