Move file reading functionality from parse_osmium_t to osmdata_t
This leaves only the functionality as Osmium handler in parse_osmium_t.

This commit also leverages the ability of osmium::io::Reader to read
from either a file or a buffer when it is called with an
osmium::io::File parameter. By creating the osmium::io::File outside the
process_file() function, we don't need the special class test_parse_t
any more.