Refactor projection code
Move away coordinates handling from functions with separate lon/lat
parameters to use of osmium::geom::Coordinates. This also removes in/out
parameters from transformation functions.

This also moves the reprojection::coords_to_tile() function as a free
function into expire-tiles.cpp, because it is only used and fits better
there. (The function is not dependent on the type of projection.)
Further refactoring is needed in the expire-tiles.cpp code.