Make clustering by geometry optional in flex output
Set table option cluster='no' to disable it. The default is
cluster='auto'. As before this does clustering with the "best" method
depending on the PostgreSQL version. Other values might be available
in the future. (If the table doesn't have a geometry column, there
is no clustering, of course.)

See #27