Remove stage 2 processing of relations
We currently have no use case for processing relations again in stage 2
and two-stage-processing is difficult enough without having to think
about how relations in relations are handled, especially when updates
are involved.

This commit removes tracking of relation ids for stage 2, i.e. the
Lua function mark_relation() and all the internal processing behind it
is removed. This functionality can be added back in later, when we
have some more experience with stage 2 processing.