Fix: Results of tasks run in thread pool were not checked
If a function run in the thread pool throws an exception, this exception
was never "collected", it was silently ignored. This commit fixes this
by wrapping the osmium::thread::Pool class in a class of our own that
keeps track of the futures returned by submit() and allows checking
them.