Extended schema support to middle and pgsql/multi output
There was already support in progress for schemas in the flex backend,
which is likely to replace the pgsql backend at some point. Until then,
it's useful to be able to import data directly into a schema.

This is the original PR from Matt Amos with some changes from me: There
are two different options now for setting the schema for the middle and
the pgsql output, respectively. There are some docs and tests now.