test-parse-osmium: do not call osmdata.stop()
The test only checks the parsing stage and does not use a database
at all. Calling stop() will try to open a database connection for
the postprocessing. This does odd things because the connection
string has not been set up. We don't need stop() for the test,
so just drop it and mark the test as NoDB.