Set projection in flex output in column config
The projection set on the command line with -l|--latlong, -m|--merc, or
-E|--proj is now ignored in the flex output. Instead you can set it
individually on all geometry columns using the `projection` option which
either takes an EPSG code or a few commonly understood string values
such as "merc", "latlong", or "WGS84".

This also makes the flex output ignore the --reproject-area option.
The area is now calculated in web mercator by default and you can
change this by setting the `projection` option on the area column.
(This currently only allows setting 4326 and 3857, other projections
will not work. We have to fix this later. This is due to the way
the area is calculated from the WKB geometry.)

The osm2pgsql.srid variable is not set any more in the Lua config.

So there are some breaking changes here for the flex output, which
we can still do as it is marked experimental.

With this change the command line options -l, -m, -E, and
--reproject-area, that can be regarded as belonging to the style are not
used any more in the flex output. This completes the move away from
command line option to config file for the flex output.