Keep the man page in markdown format and compile it with pandoc
This makes the man page easier to update and it can be included more
easily into the documentation on the osm2pgsql website.

The man page now lives in docs/osm2pgsql.md and is converted using "make
man". This uses pandoc which is readily available as packages on all
major Linux distributions and on macOS to compile the man page into the
special format for man pages.

The compiled version is still checked into the repository at the old
location docs/osm2pgsql.1 to make it easier for users.

This also installs pandoc in Linux and macOS CI builds and compiles the
man page there.