Consolidate pgsql function and trigger creation in a function
Move the code creating the pgsql function and trigger checking the
geometry validity into its own function. This function can create
the pgsql function in the correct schema if one is specified.

The flex output already has schema support, so it uses this function now
to create the function in the right schema. The pgsql output doesn't
have schema support yet, so it doesn't.