Add --with-forward-dependencies option
Add new command line option --with-forward-dependencies. Set this to
"false" to suppress the normal propagation of changes in append mode
from nodes to ways to relations. This is intended mainly for Nominatim
which doesn't need this propagation.

I have chosen the flag name as --with-forward-dependencies=VALUE to
leave the possibility open to later allow values like "nodes-only" or
something like that. But currently only "true" and "false" are
supported.

This replaces the hardcoded check for the gazeteer output in the middle.
As the (now deleted) comment in the code shows an earlier intention was
to let the output tell the middle whether it needs to handle the forward
dependencies. But this commit is not going that route: Some outputs can
not know whether the dependency management is needed (i.e. the flex
output) without having an extra flag for that. But if we need a flag
anyway, it makes more sense to have it as an overall option.

Fixes #1304