Optionally wrap polygon geometries in multipolygons
When you have an area table in the flex output, you always had to use
the "geometry" column type for the geometry column, because ways are
turned into polygons and multipolygon relations are turned into polygons
or multipolygons. The common type for these is the generic "geometry".

With this change, declaring a geometry column as type "multipolygon" now
works. All polygon geometries are automatically wrapped in a
multipolygon before adding them to the database. Depending on what you
are doing this might be want you want or not. It will make the
geometries larger, but it also makes the data more consistent and so
easier to work with and some software can only handle this consistent
data.

See #1316