Improve handling of multiple input files
Osm2pgsql can handle any number of input files. The old code will just
read the files one after the other which will not work if there is any
overlap between the files, i.e. if the same object is in two input
files.

The new code will read the files in parallel. We construct a priority
queue feeding in the next objects from all input files, taking off the
"smallest" one by one. If the same object is in multiple files, we
only process it once.

If there is only a single input file a shortcut is taken which basically
behaves like the old code.

Note that the input files have to be from the same point in time. If
there are multiple versions of the same object in the input, this will
still not magically work.

This commit removes support for unsorted input files which were already
deprecated.

See #1167

This commit removes support for negative ids which were already
deprecated.

Fixes #1097