Add support for projection using PROJ library >6.1
This commit is adding projection support using the PROJ library from
6.1 onwards.

The CMake config will prefer using the old API, simply because I don't
know how to check for the new API without it trying to use the new API
with version 5-6.0 which don't have everything we need. The code will
only work from 6.1 onwards.

There is a new CMake cache variable USE_PROJ_LIB which can be set to:
* "4": Use version 4 API. Fail CMake if it is not available.
* "6": Use version 6 API. Fail CMake if it is not available.
* "off": Disable PROJ support (resulting binary will only have WGS84
  (4326) and Web Mercator (3857) support)
* "auto": Use version 4 API if available, otherwise try version 6 API
  or fall back to no PROJ support.

Note that we have an issue with multithreading using the PROJ library,
because we use it potentially from multiple threads. The information I
can find about this seems to indicate that this is only a problem for
error reporting, because of the error handling using a global errno-type
variable. So we can't detect failed transformations reliably. But we
don't care about those anyway in the code we have, so I believe this is
probably not a huge problem. This affects the old and the new code, so
this isn't something new. The API version 6 allows to do the
multithreading correctly using a "context" which this commit does use.
But the context is still shared between threads, because we are starting
the threads after the reprojections have been initialized. So all of
this is technically not correct, but we'll probably get away with it for
the time being. We need larger refactorings in the geometry code anyway
which we want to tackle soon and hopefully can resolve this correctly
then.

See #922