Change where progress meter is updated for nodes
This is more in line with what happens for the ways and relations,
specifically the progress meter is also updated for deleted nodes
now. We also update it for all nodes, even if they are outside the
bounding box, which might or might not be correct depending on how
you interpret those numbers.