Refactor of middle: Remove slim middle class and change middle API
This commit changes the way the middle is called. The main interface for
feeding data to the middle are now the functions node/way/relation which
are called for every object, regardless of whether it is new, changed,
or deleted. The middle can then figure out itself how to handle these.
This will allow for possible optimizations when the middle can handle
an update better than a delete and create.

This also makes it natural to remove the slim_middle_t class, its
functionality is now rolled into the middle_t class. This makes the
generic middle API independent of whether we are using slim mode or not,
allowing middles which behave as slim or not depending on their own
criteria.

More refactoring based on this in later commits.