Refactor middle: Replace analyze and flush methods
Instead of the analyze() and flush() functions on the middle, the more
generic functions after_nodes(), after_ways(), and after_relations() are
introduced. Each middle can do in them whatever it needs at that place
of reading the inputs. In the case of the middle-pgsql we'll now do the
analyze of each of the nodes, ways, and relations table right after they
have been filled.

This also fixes a bug: The analyze() functions in the middle were never
called.