Remove role list parameter from rel_way_members_get()
It should not be the job of the middle to extract the roles from the
relation it retrieves. This removes the role list parameter from
rel_way_members_get() function.

But a quirk of the pgsql output is that the roles of all available
member ways of a relation are available in the Lua
"filter_tags_relation_member" callback function. This is probably not
used very often (if at all), but we need to keep that functionality
available for backwards compatibility.

So this adds a new function get_rolelist() to the pgsql output which
creates this role list.