Pull some code out of main so destructors are called earlier
This commit pulls some code creating complex objects out of main. This
way the destructors are called at the end of the run() function and
memory is cleaned up and threads are all closed. This way the final
memory and timing output will happen after everything is shut down
properly.