Use double quote on all table names etc. in middle
PostgreSQL handles plain tables name differently than it handles quoted
names. So `CREATE TABLE DEU_nodes;` is different than `CREATE TABLE
"DEU_nodes";`. In the first case the name is canonilcalized to lower
case. This quotes names in all places so we get exactly the name we ask
for.

Fixes #1420