Refactoring: Move code out of ewkb:writer_t class
Move some code out of the ewkb:writer_t class that works well as
standalone functions. This also changes the srid parameters to
use the more (for wkb) correct type uint32_t.

We also change some of the code to use the convention that pointers are
used for out parameters.